get student by ID
