+++ src/main/java/org/jdsystem/studentinfobackend/controller/StudentController.java
@@ -1,5 +1,44 @@
 package org.jdsystem.studentinfobackend.controller;
 
-public class StudentController {
 
+import lombok.RequiredArgsConstructor;
+import org.jdsystem.studentinfobackend.model.Student;
+import org.jdsystem.studentinfobackend.service.IStudentService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/students")
+@RequiredArgsConstructor
+public class StudentController {
+    private final IStudentService studentService;
+
+//    @GetMapping
+//    public ResponseEntity<List<Student> >getStudents(){
+//        return new ResponseEntity<>(studentService.getStudents(), HttpStatus.FOUND);
+//    }
+
+
+//    @PostMapping
+//    public Student addStudent(@RequestBody  Student student){
+//        return studentService.addStudent(student);
+//    }
+//
+//    @PutMapping("/update/{id}")
+//    public Student updateStudent(@RequestBody  Student student, @PathVariable Long id){
+//        return studentService.updateStudent(student, id);
+//    }
+
+//    @DeleteMapping("/delete/{id}")
+//    public void deleteStudent(@PathVariable Long id){
+//        studentService.deleteStudent(id);
+//    }
+
+    @GetMapping("/student/{id}")
+    public Student getStudentById(@PathVariable Long id){
+        return studentService.getStudentById(id);
+    }
 }
