+++ src/main/java/org/jdsystem/studentinfobackend/model/Student.java
@@ -4,16 +4,15 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import org.hibernate.annotations.NaturalId;
+import org.springframework.stereotype.Service;
 
 @Entity
-@Data
+@Getter
+@Setter
 @AllArgsConstructor
 @NoArgsConstructor
-//1231
 public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
