+++ src/main/java/org/jdsystem/studentinfobackend/repository/StudentRepository.java
@@ -3,6 +3,9 @@ package org.jdsystem.studentinfobackend.repository;
 import org.jdsystem.studentinfobackend.model.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface StudentRepository extends JpaRepository<Student,Long> {
 
+    Optional<Student> findByEmail(String email);
 }
