+++ src/main/java/org/jdsystem/studentinfobackend/service/IStudentService.java
@@ -1,0 +1,15 @@
+package org.jdsystem.studentinfobackend.service;
+
+import org.jdsystem.studentinfobackend.model.Student;
+
+import java.util.List;
+
+public interface IStudentService {
+//    List<Student> getStudents();
+//    Student addStudent(Student student);
+//    Student updateStudent(Student student, Long id);
+    Student getStudentById(Long id);
+//    void deleteStudent(Long id);
+
+
+}
