+++ src/main/resources/application.yml
@@ -10,5 +10,5 @@ spring:
   jpa:
     show-sql: true
     hibernate:
-      ddl-auto: create
+      ddl-auto: update
       format_sql: true
